chore(queue): remove unused base class
